+++ dns/doq.go
@@ -1,5 +1,3 @@
-//go:build !no_doq
-
 package dns
 
 import (
