+++ listener/tun/device/device_no_gvisor.go
@@ -1,15 +1,10 @@
-//go:build !no_gvisor
+//go:build no_gvisor
 
 package device
-
-import (
-	"gvisor.dev/gvisor/pkg/tcpip/stack"
-)
 
 // Device is the interface that implemented by network layer devices (e.g. tun),
 // and easy to use as stack.LinkEndpoint.
 type Device interface {
-	stack.LinkEndpoint
 
 	// Name returns the current name of the device.
 	Name() string
