+++ listener/tun/device/fdbased/fd_unix_gvisor.go
@@ -1,0 +1,17 @@
+//go:build !no_gvisor
+
+package fdbased
+
+import (
+	"gvisor.dev/gvisor/pkg/tcpip/stack"
+	"os"
+)
+
+type FD struct {
+	stack.LinkEndpoint
+
+	fd  int
+	mtu uint32
+
+	file *os.File
+}
