+++ listener/tun/device/fdbased/fd_unix_no_gvisor.go
@@ -1,0 +1,14 @@
+//go:build no_gvisor
+
+package fdbased
+
+import (
+	"os"
+)
+
+type FD struct {
+	fd  int
+	mtu uint32
+
+	file *os.File
+}
