+++ listener/tun/device/fdbased/open_others_no_gvisor.go
@@ -1,0 +1,11 @@
+//go:build no_gvisor && !linux && !windows
+
+package fdbased
+
+import (
+	"fmt"
+)
+
+func newEp(f *FD) error {
+	return fmt.Errorf("unsupported gvisor on the build")
+}
