+++ listener/tun/device/tun/tun_wireguard.go
@@ -8,21 +8,9 @@ import (
 	"runtime"
 
 	"github.com/Dreamacro/clash/listener/tun/device"
-	"github.com/Dreamacro/clash/listener/tun/device/iobased"
 
 	"golang.zx2c4.com/wireguard/tun"
 )
-
-type TUN struct {
-	*iobased.Endpoint
-
-	nt     *tun.NativeTun
-	mtu    uint32
-	name   string
-	offset int
-
-	cache []byte
-}
 
 func Open(name string, mtu uint32) (_ device.Device, err error) {
 	defer func() {
@@ -91,11 +79,7 @@ func (t *TUN) Write(packet []byte) (int, error) {
 }
 
 func (t *TUN) Close() error {
-	defer func(ep *iobased.Endpoint) {
-		if ep != nil {
-			ep.Close()
-		}
-	}(t.Endpoint)
+	defer closeIO(t)
 	return t.nt.Close()
 }
 
@@ -105,12 +89,7 @@ func (t *TUN) Name() string {
 }
 
 func (t *TUN) UseEndpoint() error {
-	ep, err := iobased.New(t, t.mtu, t.offset)
-	if err != nil {
-		return fmt.Errorf("create endpoint: %w", err)
-	}
-	t.Endpoint = ep
-	return nil
+	return newEq(t)
 }
 
 func (t *TUN) UseIOBased() error {
