+++ listener/tun/ipstack/gvisor/adapter/adapter.go
@@ -1,3 +1,5 @@
+//go:build !no_gvisor
+
 package adapter
 
 import (
