+++ listener/tun/ipstack/gvisor/option/option.go
@@ -1,3 +1,5 @@
+//go:build !no_gvisor
+
 package option
 
 import (
